// src/types/SqliteCard.type.ts
export interface Card {
  uuid: string;
  name: string;
  artist: string | null;
  artistIds: string | null;
  asciiName: string | null;
  attractionLights: string | null;
  availability: string | null;
  boosterTypes: string | null;
  borderColor: string | null;
  cardParts: string | null;
  colorIdentity: string | null;
  colorIndicator: string | null;
  colors: string | null;
  defense: string | null;
  duelDeck: string | null;
  edhrecRank: number | null;
  edhrecSaltiness: number | null;
  faceConvertedManaCost: number | null;
  faceFlavorName: string | null;
  faceManaValue: number | null;
  faceName: string | null;
  finishes: string | null;
  flavorName: string | null;
  flavorText: string | null;
  frameEffects: string | null;
  frameVersion: string | null;
  hand: string | null;
  hasAlternativeDeckLimit: boolean | null;
  hasContentWarning: boolean | null;
  hasFoil: boolean | null;
  hasNonFoil: boolean | null;
  isAlternative: boolean | null;
  isFullArt: boolean | null;
  isFunny: boolean | null;
  isGameChanger: boolean | null;
  isOnlineOnly: boolean | null;
  isOversized: boolean | null;
  isPromo: boolean | null;
  isRebalanced: boolean | null;
  isReprint: boolean | null;
  isReserved: boolean | null;
  isStarter: boolean | null;
  isStorySpotlight: boolean | null;
  isTextless: boolean | null;
  isTimeshifted: boolean | null;
  keywords: string | null;
  language: string | null;
  layout: string | null;
  leadershipSkills: string | null;
  life: string | null;
  loyalty: string | null;
  manaCost: string | null;
  manaValue: number | null;
  number: string | null;
  originalPrintings: string | null;
  originalReleaseDate: string | null;
  originalText: string | null;
  originalType: string | null;
  otherFaceIds: string | null;
  power: string | null;
  printings: string | null;
  promoTypes: string | null;
  rarity: string | null;
  rebalancedPrintings: string | null;
  relatedCards: string | null;
  securityStamp: string | null;
  setCode: string | null;
  side: string | null;
  signature: string | null;
  sourceProducts: string | null;
  subsets: string | null;
  subtypes: string | null;
  supertypes: string | null;
  text: string | null;
  toughness: string | null;
  type: string | null;
  types: string | null;
  variations: string | null;
  watermark: string | null;
}